server-python/requirements.txt
rembg==2.0.50
pillow==10.1.0
flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
scipy==1.11.4

